refactor: Notion API 사용하지 않도록 변경
